requirements.txt
aioredis==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
certifi==2023.11.17
cffi==1.16.0
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.1
daphne==3.0.2
defusedxml==0.8.0rc2
Django==4.2.9
django-cors-headers==4.3.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
hiredis==2.3.2
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.7
oauthlib==3.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.4
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==24.1.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
zope.interface==6.1
